vote-interface: extract `vote_state_deserialize` mod
