+++ vote-interface/src/state/mod.rs
@@ -27,6 +27,8 @@ pub mod vote_state_v4;
 pub use vote_state_v4::VoteStateV4;
 mod vote_instruction_data;
 pub use vote_instruction_data::*;
+#[cfg(any(target_os = "solana", feature = "bincode"))]
+pub(crate) mod vote_state_deserialize;
 
 /// Size of a BLS public key in a compressed point representation
 pub const BLS_PUBLIC_KEY_COMPRESSED_SIZE: usize = 48;
